+++ lucen/__init__.py
@@ -17,7 +17,7 @@ from lucen.support.errors import (
     set_errors_mode,
 )
 
-__version__ = "1.0.0"
+__version__ = "1.1.0"
 
 __all__ = [
     "activate",